% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{cnd_is}
\alias{cnd_is}
\alias{is_condition}
\alias{is_cnd_condition}
\alias{is_cnd_generator}
\alias{is_cnd_function}
\alias{is_conditioned_function}
\title{\code{is} functions for \link{cnd}}
\usage{
is_condition(x)

is_cnd_condition(x)

is_cnd_generator(x)

is_cnd_function(x, type = c("error", "warning", "message", "condition"))

is_conditioned_function(x)
}
\arguments{
\item{x}{An object}

\item{type}{A specific type to check}
}
\description{
\code{is} functions for \link{cnd}
}
