% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cndAnalysis.R
\name{cndAnalysis}
\alias{cndAnalysis}
\title{Analyse a data set from a norm}
\usage{
cndAnalysis(transfData, cndMethod, cndNorm)
}
\arguments{
\item{transfData}{an object of class \code{CndData}, with \code{X} already transformed.}

\item{cndMethod}{an object of class \code{CndMethod}.}

\item{cndNorm}{an object of class \code{CndNorm}.}
}
\value{
Return a \code{CndDatAugmented} object corresponding to \code{transfData} augmented to also include the distance and the results of the supplementary analysis.
}
\description{
Compute the distance on a new dataset, using a norm, and perform supplementary analysis.
}
\details{
The slots \code{distance} of \code{cndMethod} make use of the \code{cndNorm} to compute distance of each observation of \code{transfData} from the centroid. The result is joined to \code{transfData}, usually as a column named \code{distance} under the slot \code{other}. The new object is then used to perform other analysis, specified by the slot \code{analysis} of \code{cndMethod}, which also append the result to \code{transfData}, usually in the \code{suppl} slot.
}
\examples{
#Generate an object of class CndData.
n <- 50
yield <- 100+rnorm(n)
X <- data.frame(x1 = runif(n), x2 = runif(n), x3 = runif(n))
cndData <- CndData(yield = yield, X = X)

#Compute a reference
cndMethod <- cndMethodLandry(dropNutrient = "x3")
transfData <- cndTransform(cndData, cndMethod)
cndReference <- cndReference(transfData, cndMethod)

#Perform the analysis
cndAnalysis(transfData, cndMethod, cndNorm = cndReference@norm)

}
