% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cndAugment.R
\name{cndAugment}
\alias{cndAugment}
\title{Augment data}
\usage{
cndAugment(cndData, other = data.frame(), suppl = list())
}
\arguments{
\item{cndData}{an object of class \code{CndData} or \code{CndDataAugmented}.}

\item{other}{a \code{data.frame} of additional values to associate with points of \code{cndData}, it must possess the same number of lines than \code{yield} and \code{X}.}

\item{suppl}{\code{list} of supplementary information to pass along the object.}
}
\value{
Return an object of class \code{CndDataAugmented} with the entries \code{other} and \code{suppl} added to the corresponding slots.
}
\description{
Join additional data to a \code{CndData} object under new slots. Information as data.frame, with the same number of row as slots of \code{CndData} can be added under the slot \code{other}, while supplementary data with a different structure must be added in the \code{suppl}.
}
\details{
If \code{cndData} is already an object of class \code{CndDataAugmented}, the entry \code{other} is binded as a column to the existing slot's \code{data.frame}, while the entry \code{suppl} is combined to the corresponding slot \code{list}.
}
\examples{
#General example
##Generate an object of class CndData
n <- 20
yield <- rnorm(n)
X <- data.frame(x1 = runif(n), x2 = runif(n), x3 = runif(n))
label <- data.frame(label1 = LETTERS[1:n], label2 = 1:n)
data <- CndData(yield = yield, X = X, label = label)

##Augment data of class CndData
dataAugmented <- cndAugment(data, other = X, suppl = list(comment = "comment"))
dataAugmented

##Augment data of class CndDataAugmented
cndAugment(dataAugmented, other = X, suppl = list(comment2 = "comment2"))

}
