% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cndCateNelson.R
\name{cndCateNelson}
\alias{cndCateNelson}
\title{Cate Nelson analysis}
\usage{
cndCateNelson(cndData, n_group = 2, trend = "negative",
  labelName = NULL, x_lab = "Squared distance", y_lab = "Yield", ...)
}
\arguments{
\item{cndData}{An object of class \code{CndDataAugmented}, with a column named \code{distance} under the slot \code{other}, representing the squared distance from the centroid.}

\item{n_group, trend, x_lab, y_lab}{\code{cate_nelson}'s parameter with imposed default values specific to cnd.}

\item{labelName}{The column name of the  \code{cndData}'s \code{label} slot to be use as the argument label in the function \code{cate_nelson}.}

\item{...}{Other arguments to pass to the function \code{cate_nelson}.}
}
\description{
Perform a Cate Nelson analysis on a \code{CndDataAugmented} object. The function is a wrapper of the function \code{cate_nelson} from the \code{catenelson} package. The \code{yield} slot is used as \code{y}, and the column \code{distance} of the slot \code{other} is used as \code{x}.
}
\examples{
#General data
##CndData
n <- 20
yield <- 100+rnorm(n)
X <- data.frame(x1 = runif(n), x2 = runif(n), x3 = runif(n))
label <- LETTERS[(seq_len(n)-1)\%\%4+1]
cndData <- CndData(yield = yield, X = X, label = label)
transfData <- cndIlr(cndData)
##CndNorm and distance (CndDataAugmented)
cndNorm <- cndMcd(transfData, nSamp = 1000)
distance <- cndMahalanobis(transfData, cndNorm)

##Perform a Cate-Nelson analysis
cndCateNelson(distance, label = "label")

}
