% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cndCompositions.R
\name{cndCompositions}
\alias{cndCompositions}
\alias{cndClr}
\alias{cndIlr}
\alias{cndAlr}
\title{Transformations from the package compositions}
\usage{
cndClr(cndData, dropNutrient = NULL, ...)

cndIlr(cndData, ...)

cndAlr(cndData, ...)
}
\arguments{
\item{cndData}{an object of class \code{CndData}.}

\item{dropNutrient}{\code{character}, the name of the nutrient (column name of the slot \code{X} of \code{CndData}) to drop after the transformation.}

\item{...}{other parameters to pass to the original function of the package \code{compositions}.}
}
\value{
All functions return a transformed \code{CndData} object.
}
\description{
Wrapper of transformations \code{clr}, \code{ilr} and \code{alr} from the package compositions, to take \code{CndData} object as entry.
}
\details{
Only the requiered functions are called when needed, using the double colon symbol, to avoid namespace conflicts. The argument \code{dropNutrient} in \code{cndClr} has been added from the original function \code{clr}.
}
\examples{
#General data
##CndData
n <- 20
yield <- 100 + rnorm(n)
X <- data.frame(x1 = runif(n), x2 = runif(n), x3 = runif(n))
label <- LETTERS[(seq_len(n)-1)\%\%4+1]
cndData <- CndData(yield = yield, X = X, label = label)

#Transform the compositional component of the data,
cndIlr(cndData)
cndClr(cndData, dropNutrient = "x3")
cndAlr(cndData)
}
