% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cndCutoff.R
\name{cndCutoff}
\alias{cndCutoff}
\title{Yield cutoff subsetting}
\usage{
cndCutoff(cndData, method, param)
}
\arguments{
\item{cndData}{an object of class \code{CndData}.}

\item{method}{\code{character}, defining the basis on which to subset data. Either \code{percent} or \code{value}.}

\item{param}{\code{numeric}, specifying either the percentage of observertaion to keep (if \code{method = percent}), or the yield value delimiting the high yield subpopulation (if \code{method = value}).}
}
\value{
Return a subsetted \code{cndData} object.
}
\description{
Subset cndData on the basis of a minimum yield value, either absolute or as a percentage.
}
\details{
Only the number of observation encompassed by the percentage would be considered.
For example, with 10 observations and a percentage to preserve of 25 percent, only the first two highest yield observations would be kept.
For both method, equality is considered as part of the high yield subset (e.g. observation with a yield of 10000 would be kept in the subset if the cutoff value is also 10000).
}
\examples{
#General data
##CndData
n <- 20
yield <- 100 + rnorm(n)
X <- data.frame(x1 = runif(n), x2 = runif(n), x3 = runif(n))
label <- LETTERS[(seq_len(n)-1)\%\%4+1]
cndData <- CndData(yield = yield, X = X, label = label)

##Cutoff the data according to yield
cndCutoff(cndData, method = "percent", param = 50)

}
