% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cndMahalanobis.R
\name{cndMahalanobis}
\alias{cndMahalanobis}
\title{Mahalanobis distance}
\usage{
cndMahalanobis(cndData, cndNorm, ...)
}
\arguments{
\item{cndData}{an object of class \code{CndData}.}

\item{cndNorm}{an object of class \code{CndNorm}.}

\item{...}{other parameters to pass to the function \code{mahalanobis}.}
}
\value{
Return \code{cndData} augmented with the squared distance (\code{distance2}) as a data.frame column under the slot \code{other} (i.e. the returned object is of class CndDataAugmented).
}
\description{
Compute the Mahalanobis squared distance on the \code{X} slot of \code{cndData}, using \code{location} and \code{scatter} of \code{cndNorm}. The function is a wrapper of the function \code{mahalanobis} from the package \code{stats}.
}
\examples{
#General data
##CndData
n <- 50L
yield <- 100+rnorm(n)
X <- data.frame(x1 = runif(n), x2 = runif(n), x3 = runif(n))
label <- LETTERS[(seq_len(n)-1)\%\%4+1]
cndData <- CndData(yield = yield, X = X, label = label)
transfData <- cndIlr(cndData)

##CndNorm
cndNorm <- cndMcd(transfData, nSamp = 1000)

#Computing the distance
cndMahalanobis(transfData, cndNorm)

}
