% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cndMcd.R
\name{cndMcd}
\alias{cndMcd}
\title{Minimum covariance determinant}
\usage{
cndMcd(cndData, nSamp, ...)
}
\arguments{
\item{cndData}{an object of class \code{CndData}.}

\item{nSamp}{the number of subset used for initial estimates (\code{integer}) or the name of one of the method: \code{best}, \code{exact}, \code{deterministic}. See the function \code{covMcd} for more details on these methods.}

\item{...}{other paramaters to pass to the function \code{covMcd} from the package \code{robustbase}.}
}
\value{
Return an object of class \code{McdNorm} that contains the same slots as \code{CndNorm}, but also the logarithm of the covariance matrix determinant (slot \code{logDet}) for the best subset (slot \code{subset}, returned as an object of class \code{CndData}).
}
\description{
Compute a robust norm (location and scatter) estimate via the minimum covariance determinant (MCD), on the \code{X} component of \code{cndData}. The function is a wrapper of the function \code{covMcd} from the package \code{robustbase}.
}
\examples{
#Generate an object of class CndData.
n <- 20
yield <- 100+rnorm(n)
X <- data.frame(x1 = runif(n), x2 = runif(n), x3 = runif(n))
cndData <- CndData(yield = yield, X = X)
transfData <- cndClr(cndData, dropNutrient = "x3")

#Compute the MCD norm
cndMcd(transfData, nSamp = 1000)

}
