% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cndReference.R
\name{cndReference}
\alias{cndReference}
\title{Estimate the reference}
\usage{
cndReference(transfData, cndMethod)
}
\arguments{
\item{transfData}{an object of class \code{CndData}, with \code{X} already transformed.}

\item{cndMethod}{an object of class \code{CndMethod}.}
}
\value{
Return a \code{reference} object that contains the subset (a \code{CndDataAugmented} object that is also containing the distance) and the norm (of class \code{CndNorm}).
}
\description{
Generate a reference object, applying the \code{subset}, \code{norm} and \code{distance} components of the method, that can be used for analysis on other data sets.
}
\details{
First only preserve a \code{subset} of the data, then calculate the \code{norm} and compute the \code{distance} for this subset. Each step is specified by \code{cndMethod}.
}
\examples{
#General example
##Generate an object of class CndData
n <- 50
yield <- 100 + rnorm(n)
X <- data.frame(x1 = runif(n), x2 = runif(n), x3 = runif(n))
cndData <- CndData(yield = yield, X = X)

##Identify the method and transform the data
cndMethod <- cndMethodLandry(dropNutrient = "x3")
transfData <- cndTransform(cndData, cndMethod)

##Compute the reference according to the defined method
cndReference(transfData, cndMethod)

}
