% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cndSubsetData.R
\name{cndSubsetData}
\alias{cndSubsetData}
\title{Subset CndData}
\usage{
cndSubsetData(cndData, subset)
}
\arguments{
\item{cndData}{an object of class \code{CndData} or \code{CndDataAugmented}.}

\item{subset}{\code{logical} indicating which row to keep.}
}
\description{
Subset the slots \code{yield}, \code{X} and \code{label} of a \code{CndData} object, as well as the slot \code{other} for a \code{CndDataAugmented} object.
}
\examples{
#General example
##Generate an object of class CndData and CndDataAugmented
n <- 20
yield <- rnorm(n)
X <- data.frame(x1 = runif(n), x2 = runif(n), x3 = runif(n))
data <- CndData(yield = yield, X = X)
dataAugmented <- cndAugment(data, other = X, suppl = list(comment = "comment"))

##Subset both data sets
subset <- sample(c(TRUE,FALSE), n, replace = TRUE)
cndSubsetData(data, subset)
cndSubsetData(dataAugmented, subset)

}
