% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cndTransform.R
\name{cndTransform}
\alias{cndTransform}
\title{Transform CndData}
\usage{
cndTransform(cndData, cndMethod)
}
\arguments{
\item{cndData}{an object of class \code{CndData}.}

\item{cndMethod}{an object of class \code{CndMethod}.}
}
\value{
Return a \code{CndData} object of the transformed data.
}
\description{
Perform the \code{transformation}, specified by \code{cndMethod}, on the \code{X} component of the \code{CndData} object.
}
\examples{
#Generate random data for the example
n <- 50
yield <- 100 + rnorm(n)
X <- data.frame(x1 = runif(n), x2 = runif(n), x3 = runif(n))
label <- LETTERS[(seq_len(n)-1)\%\%4+1] #Alternative for black and white: label <- NULL

#Generate an object of class CndData.
cndData <- CndData(yield = yield, X = X, label = label)

#Generate an object of class CndMethod.
cndMethod <- cndMethodRobustIrda(dropNutrient = "x3", labelName = "label")

#Transform the composition (X) of cndData the using cndMethod
cndTransform(cndData, cndMethod)

}
