% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register.R
\name{register_conditions}
\alias{register_conditions}
\title{Register conditions}
\usage{
register_conditions(env = parent.frame())
}
\arguments{
\item{pkg}{The package name}
}
\value{
Nothing, called for its side-effects
}
\description{
\verb{[register_conditions()]} should be used within a package's \code{.onLoad()}
function to modify functions to include conditions
}
