% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4Class-CndCall.R
\docType{class}
\name{S4Class-CndCall}
\alias{S4Class-CndCall}
\alias{CndCall}
\alias{class:CndCall}
\alias{CndCall-class}
\title{CND Call (S4 Class)}
\usage{
CndCall(fun = character(0), args = list())
}
\description{
S4 class that contains the function name and arguments to be use in a method, see \code{CndMethod}.
}
\details{
The function \code{CndCall} should be used to build an object.
}
\section{Slots}{

\describe{
\item{\code{fun}}{the name of the function to call.}

\item{\code{args}}{a \code{list} of argument to pass to the function. Cnd objects used as parameter in \code{fun}, such as \code{cndData} and \code{cndNorm} should not be part of \code{args}.}
}}

\examples{
#Examine slots for the class
getSlots("CndCall")

#General example (illustrating the call for a transformation)
transformation <- CndCall(fun = "cndClr", args = list(dropNutrient = "x3"))

}
