% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4Class-CndData.R
\docType{class}
\name{S4Class-CndData}
\alias{S4Class-CndData}
\alias{CndData}
\alias{class:CndData}
\alias{CndData-class}
\alias{setValidity_CndData}
\title{CND Data (S4 class)}
\usage{
CndData(yield = data.frame(yield = numeric(0)), X = data.frame(),
  label = data.frame(), removeNA = FALSE, conformClass = FALSE)
}
\description{
S4 class that contains information on the data to use for the cnd analysis.
}
\details{
The validity method check if: \code{yield} only possesses one column; \code{yield}, \code{X} and \code{label} have the same number of lines; columns of \code{yield} and \code{X} are of class \code{numeric}.

The function \code{CndData} should be used to build an object. Both arguments \code{yield} and \code{label} can be provided as vectors, they will be coerced into \code{data.frame}. If provided as vector, the column name of \code{label} would be "label", and the one \code{yield} would be "yield". For the latter, the column name would also be overwriten as "yield" if the entry is a \code{data.frame}. If empty, the number of rows of \code{label} would be matched to those of \code{yield} and \code{X} at initialisation. \cr \cr
The logical parameter \code{conformClass} (\code{FALSE} by default) indicates if the slots \code{X} and \code{yield} must be automatically converted into \code{numeric} to respect the intended classes. The logical parameter \code{removeNA} (\code{FALSE} by default), indicates if rows containing \code{NA} in either \code{X} or \code{yield} must be removed; the operation is performed after class conversion from \code{conformClass = TRUE}.
}
\section{Slots}{

\describe{
\item{\code{yield}}{a \code{data.frame}, with one column, of yield associated to lines of \code{X}.
The entry can be a \code{numeric} vector if the object is built using the function \code{CndData}.}

\item{\code{X}}{a \code{data.frame} of composition, with columns corresponding to nutrients and lines to samples.}

\item{\code{label}}{(optional) \code{data.frame} of labels associated to lines of \code{X}. By default, an empty \code{data.frame} with the same number of rows as \code{yield} and \code{X}, but no columns.}
}}

\examples{
#Examine slots for the class
getSlots("CndData")

#General example
##Generate an object of class CndData
n <- 20
yield <- rnorm(n)
X <- data.frame(x1 = runif(n), x2 = runif(n), x3 = runif(n))
label <- LETTERS[(seq_len(n)-1)\%\%4+1]

##Generate an object of class CndData.
data <- CndData(yield = yield, X = X, label = label)

#Examples with zero or multiple label columns
##Zero columns
data <- CndData(yield = yield, X = X)

##Two columns
label <- data.frame(label1 = label, label2 = 1:n)
data <- CndData(yield = yield, X = X, label = label)

}
