% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4Class-CndNorm.R
\docType{class}
\name{S4Class-CndNorm}
\alias{S4Class-CndNorm}
\alias{CndNorm}
\alias{class:CndNorm}
\alias{CndNorm-class}
\alias{setValidity_CndNorm}
\title{CND Norm (S4 Class)}
\usage{
CndNorm(location = numeric(0), scatter = matrix(numeric(0), 0, 0))
}
\description{
S4 class that contains location and scatter estimates.
}
\details{
The validity method check if: \code{location} is of class \code{numeric}; \code{scatter} is a numerical matrix; the location length correspond to the number of row and column of scatter.

The function \code{CndNorm} can be used to build an object.
}
\section{Slots}{

\describe{
\item{\code{location}}{a \code{numeric}, location estimate.}

\item{\code{scatter}}{a \code{matrix}, variance or covariance matrix estimate.}
}}

\examples{
#Examine slots for the class
getSlots("CndNorm")

#Define the location and scatter
location <- 1:3
scatter <- matrix(1:9,3,3)

#Generate an object of class CndNorm
norm <- CndNorm(location = location, scatter = scatter)

}
