% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4Class-CndReference.R
\docType{class}
\name{S4Class-CndReference}
\alias{S4Class-CndReference}
\alias{CndReference}
\alias{class:CndReference}
\alias{CndReference-class}
\title{CND Reference (S4 Class)}
\description{
S4 class that contains the subset selected prior to the norm calculation as well as the norm. It might be used to perform analysis on another dataset. The function \code{CndReference} should be used to build an object.
}
\section{Slots}{

\describe{
\item{\code{subset}}{a \code{CndData} object representing a selection of data prior to the norm calculation. Does not includes further selection while calculating the norm.}

\item{\code{norm}}{a \code{CndCNorm} object associated to the subset.}
}}

\examples{
#Examine slots for the class
getSlots("CndReference")

#General example
##Generate an object of class CndData
n <- 50
yield <- 100 + rnorm(n)
X <- data.frame(x1 = runif(n), x2 = runif(n), x3 = runif(n))
label <- LETTERS[(seq_len(n)-1)\%\%4+1]
cndData <- CndData(yield = yield, X = X, label = label)

##Generate an object of class CndNorm
##Define the location and scatter
location <- 1:3
scatter <- matrix(1:9,3,3)
cndNorm <- CndNorm(location = location, scatter = scatter)

##Generate an object of class CndReference
reference <- CndReference(subset = cndData, norm = cndNorm)

}
