% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{standard-conditions}
\alias{standard-conditions}
\alias{InputError}
\alias{InputWarning}
\alias{ValueError}
\alias{ValueWarning}
\alias{MatchError}
\alias{PathOverwriteMessage}
\alias{PathDeletionMessage}
\alias{DeprecationWarning}
\alias{DefunctError}
\alias{TypeError}
\alias{ClassError}
\alias{NamespaceError}
\title{Standard conditions}
\usage{
InputError(x, ..., .call = getOption("cnd.call", TRUE))

InputWarning(x, ..., .call = getOption("cnd.call", TRUE))

ValueError(x, ..., .call = getOption("cnd.call", TRUE))

ValueWarning(x, ..., .call = getOption("cnd.call", TRUE))

MatchError(x, choices, ..., .call = getOption("cnd.call", TRUE))

PathOverwriteMessage(path, ..., .call = getOption("cnd.call", TRUE))

PathDeletionMessage(path, ..., .call = getOption("cnd.call", TRUE))

DeprecationWarning(x, replacement, ..., .call = getOption("cnd.call", TRUE))

DefunctError(x, ..., .call = getOption("cnd.call", TRUE))

TypeError(x, type, ..., .call = getOption("cnd.call", TRUE))

ClassError(x, class, ..., .call = getOption("cnd.call", TRUE))

NamespaceError(pkg, ..., .call = getOption("cnd.call", TRUE))
}
\arguments{
\item{x}{An object}

\item{...}{additional arguments passed to the condition message}

\item{.call}{condition arguments}

\item{choices}{A character vector of choices}

\item{path}{A path}

\item{replacement}{an object suggesting the replacement}

\item{type}{a type as returned by \code{\link[=typeof]{typeof()}}}

\item{class}{a class or vector of classes}

\item{pkg}{A package name}
}
\value{
A condition generator
}
\description{
A set of preset conditions
}
\examples{
assert <- function(object, is = character()) {
  nm <- as.character(substitute(object))
  cl <- class(is)
  if (!inherits(object, cl)) {
    cnd(ClassError(nm, cl))
  }
}

foo <- function(x) {
  assert(x, integer())
  invisible(x)
}

foo(1L)
try(foo(1.0))
try(foo("1"))

bar <- function(y) {
  assert(y, structure(list(), class = c("class1", "class2")))
  invisible(y)
}

try(bar(1L))
}
