% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnd-package.R
\docType{package}
\name{cnd-package}
\alias{cnd-package}
\title{cnd: Create and Register Conditions}
\description{
An interface for creating new condition generators, save generators in registries, link to functions, and generate documentation for packages.
}
\author{
\strong{Maintainer}: Jordan Mark Barbone \email{jmbarbone@gmail.com} (\href{https://orcid.org/0000-0001-9788-3628}{ORCID}) [copyright holder]

}
\keyword{internal}
