% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-doc.R
\name{cnd_document}
\alias{cnd_document}
\alias{cnd_section}
\title{Document your conditions}
\usage{
cnd_document(
  package = get_package(),
  registry = package,
  path = file.path("R", paste0(package, "-cnd-conditions.R"))
)

cnd_section(fun)
}
\arguments{
\item{package}{The package to document}

\item{registry}{The name of the registry}

\item{path}{The path to save the documentation}

\item{fun}{The name of a function}
}
\value{
A \code{character} vector`
}
\description{
Documents your \code{\link[=conditions]{conditions()}} and \code{\link[=conditions]{conditions()}}
}
