% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-doc.R
\name{cnd_document}
\alias{cnd_document}
\title{Document your conditions}
\usage{
cnd_document(package = get_package())
}
\arguments{
\item{package}{The package to document}
}
\value{
Nothing, called for its side-effects
}
\description{
Documents your \code{\link[=conditions]{conditions()}} and \code{\link[=conditions]{conditions()}}
}
