% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condition.R
\name{condition}
\alias{condition}
\alias{conditions}
\title{Conditions}
\usage{
condition(
  class,
  message = NULL,
  type = c("error", "warning", "message"),
  exports = NULL,
  help = NULL,
  package = get_package()
)

conditions(class = NULL, package = NULL)
}
\arguments{
\item{class}{The name of the new class}

\item{message}{The message to be displayed when the condition is called}

\item{type}{The type of condition: error, warning, or message}

\item{exports}{The exported functions to be displayed when the condition is
called}

\item{help}{The help message to be displayed for the condition function}

\item{package}{The package to which the condition belongs}
}
\value{
\itemize{
\item \code{\link[=condition]{condition()}} a \code{cnd::condition_function} object
\item \code{\link[=conditions]{conditions()}} a \code{list} of all conditions
}
}
\description{
\code{\link[=conditions]{conditions()}} is used to create a new condition function that itself returns
a new \code{condition}.
\code{\link[=conditions]{conditions()}} retrieves all conditions based on search values
}
\details{
Conditions
}
\examples{
# create a new condition:
cond_value_error <- condition("value_error")

# use the condition
try(stop(cond_value_error()))

# dynamic messages:
cond_class_error <- condition(
  "class_error",
  message = \(x) paste("class cannot be", toString(class(x)))
)
try(stop(cond_class_error(list())))

}
