% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{defaults}
\alias{defaults}
\alias{value_error}
\alias{class_error}
\alias{type_error}
\alias{input_error}
\alias{use_error}
\alias{duplicate_error}
\alias{defunct_error}
\alias{deprecated_warning}
\alias{value_warning}
\alias{type_warning}
\alias{input_warning}
\alias{class_warning}
\alias{use_warning}
\alias{duplicate_warning}
\title{Default conditions}
\usage{
value_error(..., .call = getOption("cnd.call", TRUE))

class_error(
  ...,
  x,
  actual = x,
  actual_class = class(actual),
  expected,
  expected_class = class(expected),
  name,
  .call = getOption("cnd.call", TRUE)
)

type_error(
  ...,
  x,
  actual = x,
  actual_type = typeof(actual),
  expected,
  expected_type = typeof(expected),
  name,
  .call = getOption("cnd.call", TRUE)
)

input_error(..., .call = getOption("cnd.call", TRUE))

use_error(..., .call = getOption("cnd.call", TRUE))

duplicate_error(
  ...,
  x,
  positions = which(duplicated(x)),
  duplicates = x[positions],
  name,
  .call = getOption("cnd.call", TRUE)
)

defunct_error(..., defunct, replacement, .call = getOption("cnd.call", TRUE))

deprecated_warning(
  ...,
  deprecated,
  replacement,
  version,
  .call = getOption("cnd.call", TRUE)
)

value_warning(..., .call = getOption("cnd.call", TRUE))

type_warning(
  ...,
  x,
  actual = x,
  actual_type = typeof(actual),
  expected,
  expected_type = typeof(expected),
  name,
  .call = getOption("cnd.call", TRUE)
)

input_warning(..., .call = getOption("cnd.call", TRUE))

class_warning(
  ...,
  x,
  actual = x,
  actual_class = class(actual),
  expected,
  expected_class = class(expected),
  name,
  .call = getOption("cnd.call", TRUE)
)

use_warning(..., .call = getOption("cnd.call", TRUE))

duplicate_warning(
  ...,
  x,
  positions = which(duplicated(x)),
  duplicates = x[positions],
  name,
  .call = getOption("cnd.call", TRUE)
)
}
\arguments{
\item{...}{Additional message components; extra named arguments will signal
an \code{\link[=input_warning]{input_warning()}}.}

\item{x}{an object}

\item{actual, expected}{Objects whose \code{class} or \code{type} should be retrieved}

\item{actual_class, actual_type, expected_class, expected_type}{Override \code{class}
or \code{type}}

\item{name}{Name of the object (will be \code{\link[base:deparse]{base::deparse()}}'d if not provided)}

\item{duplicates}{Duplicated values of \code{x}}

\item{defunct, deprecated, replacement}{Defunct, deprecated and replacement
object, use \code{\link[base:substitute]{base::quote()}} to pass expressions (e.g., \code{quote(fun(old = ))})}

\item{version}{A version number}

\item{position}{Vector positions of \code{x}}
}
\value{
A condition object
}
\description{
Default conditions have options for dynamically generated messages based on
the parameters provided.
}
\details{
If no values are entered into the \link{condition_generator}, a
default message will be used. Messages will be dynamically created based on
the parameters provided.
}
\section{messages}{
 Messages are dynamically created with specific inputs per
condition. Generally, parameters must be explicitly used.  All condition
messages use \code{...} which allow for either additional context to be added or
completely overriding the message.  Default messages will require the first
parameter to be used.
}

\examples{
type_error()
type_error(x = 1L)
type_error(x = 1L, expected = double())
type_error(x = 1, actual = integer(), expected = double())
type_error(
  "Additional context",
  x = "a",
  expected = 1L
)

}
\keyword{internal}
