% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{defaults}
\alias{defaults}
\alias{value_error}
\alias{class_error}
\alias{type_error}
\alias{input_error}
\alias{defunct_error}
\alias{deprecated_warning}
\alias{value_warning}
\alias{type_warning}
\alias{input_warning}
\alias{class_warning}
\title{Default conditions}
\usage{
value_error(..., ..., .call = getOption("cnd.call", TRUE))

class_error(
  ...,
  x,
  expected,
  actual = x,
  name,
  ...,
  .call = getOption("cnd.call", TRUE)
)

type_error(
  ...,
  x,
  expected,
  actual = x,
  name,
  ...,
  .call = getOption("cnd.call", TRUE)
)

input_error(..., ..., .call = getOption("cnd.call", TRUE))

defunct_error(
  ...,
  defunct,
  replacement,
  version,
  ...,
  .call = getOption("cnd.call", TRUE)
)

deprecated_warning(
  ...,
  deprecated,
  replacement,
  version,
  ...,
  .call = getOption("cnd.call", TRUE)
)

value_warning(..., ..., .call = getOption("cnd.call", TRUE))

type_warning(
  ...,
  x,
  expected,
  actual = x,
  name,
  ...,
  .call = getOption("cnd.call", TRUE)
)

input_warning(..., ..., .call = getOption("cnd.call", TRUE))

class_warning(
  ...,
  x,
  expected,
  actual = x,
  name,
  ...,
  .call = getOption("cnd.call", TRUE)
)
}
\arguments{
\item{...}{Additional message components}

\item{x}{an object}

\item{actual, expected}{Actual or expected value (see details)}

\item{name}{Name of the object (will be deparsed if not provided)}

\item{defunct, deprecated, replacement}{Defunct, deprecated and replacement
object, use \code{\link[base:substitute]{base::quote()}} to pass expressions (e.g., \verb{quote(old_fun()})}

\item{version}{A version number}
}
\value{
A condition object
}
\description{
Default conditions have options for dynamically generated messages based on
the parameters provided.
}
\details{
If no values are entered into the \link{condition_generator}, a
default message will be used. Messages will be dynamically created based on
the parameters provided.
}
\section{messages}{
 Messages are dynamically created with specific inputs per
condition. Generally, parameters must be explicitly used.  All condition
messages use \code{...} which allow for either additional context to be added or
completely overriding the message.  Default messages will require the first
parameter to be used.
}

\examples{
type_error()
type_error(x = 1L)
type_error(x = 1L, expected = double())
type_error(x = 1, actual = integer(), expected = double())
type_error(
  "Additional context",
  x = "a",
  actual = I("string"),
  expected = I("numeric or integer")
)

}
\keyword{internal}
