% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{format-conditions}
\alias{format-conditions}
\alias{format.cnd::condition}
\alias{format.cnd::condition_generator}
\title{Format conditions}
\usage{
\method{format}{`cnd::condition`}(x, ..., cli = getOption("cnd.cli.on", TRUE))

\method{format}{`cnd::condition_generator`}(x, ..., cli = getOption("cnd.cli.on", TRUE))
}
\arguments{
\item{x}{A \link{condition} object}

\item{...}{Not used}

\item{cli}{If \code{TRUE} will use formatting from \code{\link[=cli-package]{\{cli\}}}}
}
\description{
Formats \link{condition} objects
}
