% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{format-conditions}
\alias{format-conditions}
\alias{format.cnd::condition}
\alias{format.cnd::condition_generator}
\title{Format conditions}
\usage{
\method{format}{`cnd::condition`}(x, ..., cli = getOption("cnd.cli.override"))

\method{format}{`cnd::condition_generator`}(x, ..., cli = getOption("cnd.cli.override"))
}
\arguments{
\item{x}{A \link{condition} object}

\item{...}{Not used}

\item{cli}{If \code{TRUE} will use formatting from \link[cli:cli-package]{cli}.
Default uses an option, \code{"cnd.cli.override"}, if available, otherwise
checks that \code{cli} is installed and ansi colors are available.}
}
\value{
A \code{character} vector
}
\description{
Formats \link{condition} objects
}
\examples{
format(condition("foo"))
}
